+++ just-plan/src/app/(needLogin)/_lib/getSearchCity.ts
@@ -1,5 +1,5 @@
 import { nextFetch } from "@/lib/returnFetch";
-import { IRegion } from "@/types/plan.types";
+import type { IRegion } from "@/types/plan.types";
 
 interface IBody {
   cities: IRegion[];
@@ -7,6 +7,5 @@ interface IBody {
 
 export const getSearchCity = async (cityName: string) => {
   const res = await nextFetch(`/api/cities/search?cityName=${cityName}`);
-
   return res as unknown as IBody;
 };
